Update NormalMatrix(elements) constructor
As it is used in MatrixAdt multiplication operator when constructing
the resulting MatrixAdt that take NormalMatrix that accepts elements.

Previous implementation causes ambiguity between constructors
overloads.

The solution was to take elements as value type, this also lead to
adding a type alias and using it where it is needed,
to avoid duplication.