Fix Address Sanitized tests
There was a copy-paste error when writing the tests for
LowerTriangularMatrix that caused memory access beyond boundaries.

This issue should 've been caught at build time like NormalMatrix
implementation, but this didn't happen, probably due to several
indirection calls to initialize the matrix elements. To fix it,
the following was performed:
1- Added a static assert to the used constructor to give
   a build-time error for this problem instead
2- Fixing the size of matrices in the unit tests to pass the failing
   assertions