Document the issue faced with thread-sanitizer
So, thread sanitized unit tests were always failing with either
unexpected memory mapping error or a coredump.

It turned out it had to do with virtual memory entropy limit,
it even behaved the same with a simple hello world program.

The fix was lowering the entropy limit to 28, and the issue
no longer was observed.