Make advance() function signature more clear
Made the node argument of type std::reference_wrapper to a generic
type to accept both mutable and immutable objects, instead of just
using "auto&" where it doesn't show that reference_wrapper is
expected by this method, even though a call to .get() method is
performed inside which is provided by reference_wrappers.

This also saves an unecessary overloading of duplicating the
implementation for both const and non-const objects.