Apply Rule of five to LinkedList class
Since concatenate method implementation needs to provide explicit
implementation to the move constructor to handle unique pointers
movement properly.

Accordingly, the rest of the special member functions have to be
providedd. Where the copy constructor and assignment can be deleted
to keep uniqueness of the nodes, and the destructor can be simply
defaulted.

In addition, the move assignment operator will make use of the
move constructor using the move and swap idiom, where for such
case a swap() method was provided as well.