Optimize push() and pop() operations
Instead of pushing and popping to the last position
in the container, which could yield a O(N) time complexity
in some containers like a linked list, we can push and pop
the first position instead, which would be O(1) in this case.

Also needed to update the tests and the concept requirements
according to this API change.

Note: It's always O(1) in case of using an array